An abstract class for paginated list provider
