+++ app/src/main/java/derpibooru/derpy/server/providers/ImageListProvider.java
@@ -9,32 +9,11 @@ import derpibooru.derpy.data.server.DerpibooruTagDetailed;
 import derpibooru.derpy.server.QueryHandler;
 import derpibooru.derpy.server.parsers.ImageListParser;
 
-public class ImageListProvider extends Provider<List<DerpibooruImageThumb>> {
+public class ImageListProvider extends PaginatedListProvider<DerpibooruImageThumb> {
     protected static final int IMAGES_PER_PAGE = 16;
-
-    private int mCurrentPage = 1;
 
     public ImageListProvider(Context context, QueryHandler<List<DerpibooruImageThumb>> handler) {
         super(context, handler);
-    }
-
-    public ImageListProvider nextPage() {
-        mCurrentPage++;
-        return this;
-    }
-
-    public ImageListProvider resetPageNumber() {
-        mCurrentPage = 1;
-        return this;
-    }
-
-    public ImageListProvider fromPage(int page) {
-        mCurrentPage = page;
-        return this;
-    }
-
-    public int getCurrentPage() {
-        return mCurrentPage;
     }
 
     @Override
